+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Log dataset sizes and digest on changes.
+
 ## [2.2.1] - 2025-12-06
 
 ### Added